Moving setting position absolute on inserted element to insertmenu
